keep track of piece availability.  we simply keep an array with integers, one element per piece.  it holds the count for that piece.
when a peer sends a "bitstring" or "have" message, the piecescounts for the relevant pieces are increased by one.
when a peer quits, the "piecescount" is decreased by one for each each peer the piece had.
this should keep the account accurate at all time.

when selecting a piece, use random pieces for the first n (4 currently) pieces.  choose rarest-first for the other pieces, randomly chosing from the pieces that are equally the rarest.  note that this doesn't help much yet:  we do not yet request blocks from the same piece from different peers during startup-phase (when picking pieces at random).  using multiple peers to get a single piece would be useful.  the best chances to get the first piece is probably to pick one that not every peer already has, but that a sufficient number of peers do have, so the piece is likely to be completed quickly enough.