implement choking algorithm.  there are different algorithms for normal peer and for seed state.  the algorithm from the bittorrent spec was used, updated with information from the paper "Rarest First and Choke Algorithms Are Enough" by Legout, Urvoy-Keller and Michiardi.
for seeding, we keep 4 peers unchoked.  every 10 seconds we start a round.  every 3rd round we do nothing every 1st and 2nd round, we choke the peer that has been unchoked the longest and replace it by a random interested but choked peer.  the other 3 peers are kept unchoked.

for normal peer state, we also keep 4 peers unchoked.  again, rounds of 10 seconds are used.  every third round, we pick a new peer to unchoke (xxx the peer should be interested, but currently an uninterested peer can also be picked... (or not?)).  we then choose the four fastest (to us) peers that are interested in us.  we unchoke all peers that are uninterested but faster than the fastest peer (xxx should be the one faster than the slowest peer we'll keep unchoked.  next we unchoke the optimistic peer and up to four of the fastest peers (we only unchoke three if the optimisticly unchoked peer was interested).

the optimistic unchoke is chosen round-robin.  new connections are unchoked more often than existing connections (3:1 ratio).  the original bittorrent spec specified it so.  this should probably be changed to random (just as with the new seeding algorithm).  otherwise malicious clients can simply disconnect and reconnect to get an advantage.  this can be remedied by punishing reconnects.  but in large clouds the malicious peer will have enough preypeers to abuse.  doing it random will give abusers no advantage.  and will reduce code size.

warning:  needs a lot of testing and refinement.