write parts of a piece to disk in Diskchunksize portions.  remove the manual writing of data in Diskchunksize portions.  this prevents fragmenting as well.  previously, we kept an entire piece in memory, which will take too much memory for large files (thus large pieces).  now we'll never use more than 128k per peer, but on average, we'll use half that.
calculate the hash of piece while parts are coming in.  previously, we calculated it at the end (with the in-memory piece).  now we keep track of the hash for blocks that come in in-order.  in the end, we'll read from disk if necessary, to calculate the hash over the remainder of the data.

make bandwidth shaping simpler.  now we just give out the requested data (well, normally, fewer bytes), and sleep for the time the given bandwidth covered.  this makes us never allocate more than the maximum, and spreads the data evenly over time.  if we change the bandwidth later on, it will just work given a new value.