- use rand(2), not /dev/random (which is slower) for randomizing e.g. pieces, peer selection.  we seed rand(2) with data from /dev/random, so should be reasonably secure.
- introduce adt Rotation for rotating through lists.  this is currently used to rotate through ip addresses in seed-mode.  previously, this had local code to fill & pick from the pool.  this will probably be extended to allow random selection, in order selection and pool replenishing.

- code for sorting is merged (duplicate code reduced).  we now have one sorting function, which accepts a compare function reference as argument to do the comparing.


piece accounting has changed:

- we keep track of active pieces (those that we actively try to download).  finished pieces are not active.  inactive pieces are those that we have not yet attempted to download (e.g. because they were less rare than others).  active pieces may be "orphans":  no requests for blocks of these blocks are pending at peers.  orphan blocks usually have preference when scheduling.
- we keep track of which peers are busy on each block.  multiple peers can have requests for a block, for faster operation.  we will now send cancel messages when necessary.


scheduling has changed completely:

before this commit, we assigned a single piece to a peer.  let the peer download & verify the piece, and then continue on the next one.

after this commit, assign (batches of) blocks to peers, and assign blocks to two peers if necessary.  function schedule(peer) schedules new blocks for a peer.  this is called on every event for which we may want to send new requests to a peer (such as a peer completing a piece we are interested in).

we take into account whether we are in rarest-first or random-piece mode.  we can assign multiple peers to a piece (most notably during startup & in end-game).  until we have four complete pieces, we are in random-piece mode.  peers are assigned to the most completed pieces in order to complete a piece as quickly as possible.  during rarest-first mode, we try to assign a single peer to a piece.  of course, when a peer chokes us, such pieces become orphans and will be picked up with preference by another peer that unchokes us (to prevent many half-downloaded pieces to linger around).  if there are no more new pieces to start on, we start on pieces other peers are already working on.  if there are no more blocks that have not yet been requested, we start requesting blocks a second time (in reverse order), to make sure the download finishes faster.

we assign blocks of a piece preferably in batches.  each peer tries to finish (it's part of) the batch, and write it to disk in one go.  a batch is 128k.  this prevents too many random writes (slow & disk-fragmenting) for most of the pieces (i.e. after startup and before end-game, normally the bulk of the pieces).

scheduling will be refined.  first, the current heuristics can be refined a bit.  second, we'll change behaviour based on mode & peer reputation (keeping track of whether we are under attack, to be resilient against organisations that try to bring down a torrent).  third, there is plenty of room for optimisation (more careful block-accounting), we recalculate state too much.


caveat:  still have to test with multiple peers, so there will be fallout :)