fix downloading.  the last pieces would no longer come in because we would send an Uninterested message once all pieces were busy.
now we first check if we have requests scheduled for the peer (cheap check).  and otherwise see if there are new pieces to be requested.  the function interesting(peer) should be called more often.  we could be no longer interested in a peer that has a piece we don't.  for example, if we already requested all blocks of such pieces from too many other peers.  will be fixed later.