big changes.  mostly simplifications.
we no longer requests rare pieces first.  instead, we always request
at random.  this requires much less accounting, and is not susceptible
to all kinds of poisoning attacks.

we are also stricter with assigning pieces to peers.  this is part
of the defense mechanism against malicious peers.  it is quite hard
to infallibly detect malicious peers, so we give up on that.  the
original idea was that you could find out which peer had been giving
bogus blocks of a piece.  instead, we do mark peers as good as soon
as they have given us a piece (and we verified it) all by themselves.
we further divide pieces into three categories:  active, orphans,
inactive.  active pieces are those for which we have requests
outstanding at peers (that have us unchoked).  orphans are pieces
of which we already have at least one block and for which we have
no peer working on (no outstanding requests).  all the other pieces
are inactive pieces.

for scheduling:
- try to use the piece the peer was last working on.  if no other
  piece has worked on it since, we schedule as many blocks from it
  as possible (sequentially).
- if the peer is good, we let it continue on orphan pieces.  those
  with most blocks already downloaded first, giving us completed
  pieces as soon as possible.  non-good peers are skipped, they might
  be poisoning our partially completed block.
- next we continue with new random pieces, as many as possible.
- finally, for non-good peers, we schedule orphans to it.  if we
  wouldn't, these unproven peers would be completely useless.

note that we don't yet have an end-game.  previously, we much earlier
try to request the same block from multiple peers.  but that involves
having to send cancel messages too, which is annoying to account.
we currently don't ever send cancel messages.  we probably should
have an end-game in the future.

this also makes sure that buffers from peers that are dropped are
still flushed completely.  before, flushing was managed by
peernetreader, which was killed when the peer was dropped.  flushing
is still managed by peernetreader, but it is no longer killed:  if
the peer is dropped, the peernetreader will hear about it on a
channel, then stop itself.

also fix reading/writing for torrents with many files.

when parsing torrent files, look for "private", "created by" and
"creation date" too.

buffering data, to write in one big chunk so it is contiguous on
disk, has been clarified a bit too, with a queue and new adt's
instead of (reversed) lists of tuples.