implement an end-game.  and many other small changes.  this also adds quite some debugging messages, to be removed later.
the end-game does change quite a lot of the peer/piece/block
accounting.  we still keep active/orphan/inactive piece information.
but peers remain interested as long as there is any piece (regardless
of active/orphan/inactive) that we do not yet have.  before, we did
not want to schedule peers on active pieces.  however, this makes
last blocks go extremely slow when it is allocated to a slow (or
stalling) peer.  we still don't like assigning multiple peers on
one piece, but in general, if you want to make progress in all
cases, you will have to.  deciding when (not) to is quite hard.  in
the future, we might postpone scheduling unknown quality (or known
faulty) peers on an active/orphan piece, only scheduling such peers
on such pieces after we haven't made progress for some time.

anyway, we now have Peer.lwant for which pieces the peer has and
we want (was lwantinact).  we also have Peer.canschedule, for pieces
we can schedule on.  this is used for determining whether we are
interested in a peer.  as long as there are blocks in the piece
that we haven't requested to that peer, we remain interested and
the piece remains canschedule.

another consequence of scheduling multiple peers on a single piece
(and block) is that we now send cancel messages to all other peers
we requested a received block from.

when scheduling multiple peers on a single piece, we first requests
blocks once (whichever peer is scheduled).  then we continue
requesting blocks (that the peer hasn't requested yet, but other
peers have) randomly, until that peer has requested all blocks of
the piece.

the number of blocks we allocate depends on the download rate.  we
have a minimum of 8 requests scheduled normally.  but on higher
download rates, we schedule more requests.  the mechanism to calculate
number of requests to send needs more tweaking though.

to account for the active peers on a piece, we now have a Piecepeer
to account for peer activity on a Piece.  active pieces have
piecepeers.

be a bit less agressive with sending requests.